auto logout user successfully working
